requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
requests==2.31.0
certifi==2024.2.2
idna==3.6
python-dotenv~=1.0.1
google-auth
google~=3.0.0